requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.0.2
idna==2.7
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.10
lxml==4.2.3
MarkupSafe==1.0
newspaper3k==0.2.6
nltk==3.3.0
olefile==0.45.1
Pillow==5.2.0
py==1.5.4
pytest==3.6.3
python-dateutil==2.7.3
PyYAML==3.13
requests==2.19.1
requests-file==1.4.3
six==1.11.0
tldextract==2.2.0
urllib3==1.23
Werkzeug==0.14.1
Rx==1.6.1
elasticsearch==6.3.0
faker==0.8.16
numpy==1.14.5
